Improve popup contrast styles. (#603)
* Improve popup contrast styles.

* Improve popup contrast styles.

* Improve popup contrast styles.

* Improve popup contrast fader style.

* Refactor component style override.

* Refactor component style override.
